Consider private tags while parsing Eigen Artemis 3D US
This is a followup on #6 adding additional checks and Eigen Artemis
private dictionary.

Parsing of the volume spacing is moved from load() to examine(). This
will prevent errors on load when spacing cannot be found, and will also
avoid loading of the dataset twice, first in examine(), and again in
load().