ENH: Add ConformalTextureMapping CLI module
It uses the LSCM executable and a CLI XML descriptor to run the LSCM algorithm on a model with defined fixed points.

The only problem is that starting the CLI from Slicer it fails with a single error "Conformal Texture Mapping terminated with a fault.", but when the same command is run manually from the command-line it works well.

Re https://github.com/JolleyLabCHOP/SlicerHeartPrivate/issues/159