Allow devices to create arbitrary models
Previously, devices were always rotationally symmetric and their model was created by rotational sweep of a profile.

Now if a device implements updateModel(modelNode, parameterNode) then it can provide a model with arbitrary geometry. It is useful for devices that are not rotationally symmetric, such as MitraClip (https://github.com/JolleyLabCHOP/Internal/issues/23).

Also fixed drifting device position when repeatedly clicking on "Position device at centerline", and positioning of device near the end of centerline curve (it did not quite reach the end of the curve).