ENH: Generalize US 3D Reader to also support Image3dAPI through manifest
This commit generalize the support for Image3dFileLoader so that the
loading of the associated type library wrapper can be done if it was
directly associated with the application using manifest.

This removes the need for administrator privileges needed to explicitly
register the loader using command like "regsvr32 Image3dLoaderGe.dll".

At the time of this commit only the GE loader supports this approach.