ENH: Add FluoroSimulator module
The "Virtual Cath Lab" module simulates fluoroscopy using a single-plane or bi-plane C-arm.

Motion for each component of the C-arm can be controlled using the module. As the C-arm is moved, the simulated image will be updated in real-time. This image is generated using volume rendering on a CT image using custom transfer functions.

Users can also simulate the addition of contrast agent in the image by filling in segmented regions of the with high intensity values.