BUG: Handle fan branching on the extremes of the boundary edges
Where a primary chord needed to be branched out near one end of the boundary edge (the edge of the grid surface between the region division lines), it was possible that the edge section determined for the fan branching overreached the end of said line. It was handled in a wrapping way, i.e. the next point wrapped around, and it caused "stray" chord branches. This change makes sure that if the edge section would reach over, it instead extends to the other direction (does not stop early, just extends the other way).