BUG: Fix Philips4dUsDicomPatcher not exporting NRRD file
Examination of DICOM files broke in Slicer-5.8 for files that are not in the database
(due to slicer.dicomDatabase.fileValue() only inspecting files that are in the database).
This made isDicomUltrasoundFile() method always return false and skipping NRRD export.

Fixed by only trusting fileValue result if it returns non-empty string.